[threading] support dict (~> 'a {'a 10})
Since ~> is a macro, it is not possible to detect '{ }' as dictionaries, so
that the above form will expand to `{'a 'a 10}` and later be expanded by #%app
to `(hash 'a 'a 10)` and thow error.

So we modify ~> to treat any form other than `(...)` as data instead of
function application. So now we cannot use `[...]` or `{...}` to represent
applications inside ~>.